feat: Show Last Updated As Time Ago
Resolves #132

Parses last updated date string and displays a relative time
difference (includes i18n).

Currently expects the date string to
be in the format "MMM DD YYYY, HH:mm z" as in
"Mar 29 2020, 15:53 JST". If the API is updated to instead use
a Unix timestamp (or other directly parseable format), then the
`parseDate` function can be removed in favor of passing the
timestamp directly to the Date constructor.