Improve Mapbox loading
Change JS loading to be async.
Properly handle cases where Mapbox is not loaded and initMap() is called. (This will also fix errors for mapboxgl undefined.)

Move globals for tippy and pageDrawCounts into their own modules so they are not exposed.

Move all index.js page state into a single object so it's more readable.

If mapboxgl is not available, pause for 3 seconds before retrying.

Stop passing unnecssary arguments to drawMap() which are either not used or can be accessed globally.

Move some globals to be within the module where they are only used.